completed some exercies; started 1.33
All of the new exercises are complete except for 1.33. In 1.33, I
decided to do the assignment a little differently, explained in the
comment. The only problem with it (AFAIK) is that relatively-prime?
does not work (it gets stuck in an infinite loop on certain inputs).

I un-optimized 1.16 because I've started using it in other contexts
where we can't assume that a number is odd just because we input a
value greater than 2.