Refactored multiply function in 2.1.4
There is now a naming convention for multiply functions. For example, if
you are multiplying (-2, 4) and (2, 7) then
multiply-interval-neg-pos-pos-pos is called. The function name is
computed dynamically so that the user only ever has to call
multiply-interval.

I could do a similar thing with division, subtraction, and addition but
I feel like I've learned what I needed to from this lesson so I'm going
to move on.