Fixed eval bug in 2.11, added explanation to 2.14
I also properly renamed 2.13 to 2.14

Eval takes 2 arguments, not 1, so I added the second one to 2.11.scm.
The second argument indicates what environment to evaluate in.

Added an explanation to 2.14 about why the two expressions are not
equivalent under interval arithmetic. I also added an implementation of
add-interval to the file, as that was deleted when I split up 2.1.4.scm
into exercise files.