added 2.11 changes for real,reanamed/extended 2.16
When I fixed eval two commits ago I forgot to add the file. It's
actually in here now.

I realized that the proper name for 2.14 was *actually* 2.16, so it's
named that now. I also answered the second part of the exercise asking
if it was possible to create a "beter" system (more conformant to
integral algebra).