Fixed two bugs in the queens problem (2.42)
Both problems were with the implementation of safe?. The first problem
was that safe was returning a list of booleans instead of a single
boolean. The second problem was in the test for diagonals: it was
comparing the difference of the rows and columns being tested instead of
the absolute value of the difference. The safe? tests have been updated
to take into account both of these problems, by changing the expected
return value of the functions and splitting diagonally-unsafe-board into
descending-diagonally-unsafe-board and
ascending-diagonally-unsafe-board.